+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 ### Fixed
 - Fixed a bug where Neo input blocks could be dragged to positions where they would exceed their field's Max Levels setting
+- Fixed a bug where Neo blocks' enabled and collapsed states weren't retained in some cases when cloning them
 
 ## 2.9.6 - 2021-04-14